pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "anomrl"
version = "0.1"
readme = "README.md"
requires-python = ">=3.10,<3.11"
dependencies = [
    "av==12.1.0",
    "gymnasium==1.1.1",
    "hydra-core==1.3.2",
    "lightning==2.3.3",
    "minari[all,hdf5]==0.5.0",
    "mlflow==2.21",
    "mujoco==3.2.6",
    "numpy<2.0",
    "psutil==6.0.0",
    "pygame==2.5.2",
    "pyquaternion==0.9.9",
    "python-dotenv==1.0.1",
    "rich==13.7.1",
    "sb3_contrib==2.3.0",
    "seaborn==0.13.2",
    "stable-baselines3==2.6",
    "torch==2.4.0",
    "torchvision==0.19.0",
    "tqdm==4.66.4",
    "imageio>=2.14.1",
    "pynvml>=12.0.0",
    "dotenv>=0.9.9",
    "tables>=3.10.1",
    "anomaly-gym",
    "tsb-ad>=1.5",
]

[project.optional-dependencies]
extra = [
    "jupyter>=1.1.1",
]



[tool.ruff]
line-length = 120
indent-width = 4
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "B", "UP", "C90", "RUF"]
ignore = ["B028", "RUF013"]
fixable = ["ALL"]
unfixable = []
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["E402"]
"**/{tests,docs,tools}/*" = ["E402"]

[tool.uv.sources]
torch = [
    { index = "pytorch-cu124" },
]
torchvision = [
    { index = "pytorch-cu124" },
]
anomaly-gym = { git = "https://github.com/anonymous-3-141592653/anomaly-gym"}


[[tool.uv.index]]
name = "pytorch-cu124"
url = "https://download.pytorch.org/whl/cu124"
explicit = true
